cpio: drop hardlink tracking
This significantly simplifies archiving. Any hardlink will be treated as
a normal file: assigned a unique inode number and associated data
segment. Data segments may be duplicated if multiple hardlinks with the
same source inode are archived (easy to avoid).

The main benefit here is that deferred I/O for hardlink data segments is
removed, so src-path = archive-path behaviour can finally be dropped.

Signed-off-by: David Disseldorp <ddiss@suse.de>