cpio: accept metadata via new ArchiveMd type
The fs::Metadata type is private, which makes it difficult to create
cpio archive entries for items which don't exist locally. Create a
new public ArchiveMd type for this purpose, which can easily be
instantiated from fs::Metadata .
This also has the benefit of moving some of the duplicate 64-bit ->
32-bit (cpio) overflow checks into a single function.
There's no backward compatibility offered, so this change will break
existing callers.

Signed-off-by: David Disseldorp <ddiss@suse.de>