rs/rapido-cut: handle --install src→dest parameters
We now have the ability to mock up ArchiveMd metadata for cpio entries,
so src→dest install can be properly implemented. Support is still a bit
ugly in that symlinks are not handled - it gets a bit messy / complex
if we need to modify target paths, so I'd prefer to skip it if we can
get away with it.

Signed-off-by: David Disseldorp <ddiss@suse.de>