rs/dracut-cpio: rework tests following '/' prefix stripping
Tests which use absolute paths and compare binary output to GNU cpio
need to be reworked, such that GNU cpio also omits the '/' prefix.
This is done by changing to the root directory.

When cpio.rs test coverage improves, we can probably drop dracut-cpio.
But cpio.rs tests have a long way to go before we're at that point.

Signed-off-by: David Disseldorp <ddiss@suse.de>