rs/rapido-vm: avoid traversing full image for rapido-rsc
When determining VM cpu, memory and network resources, we don't need to
traverse the entire cpio archive if we know that all rapido-rsc entries
are located together. Assume this is the case (it is with Dracut) and
break traversal when stepping past a previously encountered rapido-rsc
path.

This might be a premature optimization, in which case I'm sorry.

Signed-off-by: David Disseldorp <ddiss@suse.de>