rs/rapido-cut: fix bin/lib symlink handling
Symlinks can carry arbitrary text. relative symlinks are supposed to be
interpreted as relative to the link parent dir, not current working dir.
Fix this by resolving the link target via canonicalize() and adding that
as gather dependency. This drops any intermediate symlinks, but that
shouldn't be a big problem.

Signed-off-by: David Disseldorp <ddiss@suse.de>